// src/stores/default/types.ts
export type ServiceInfoType = {
  service_name_ko: string;
  service_name_en: string;
  company_name_ko: string;
  company_name_en: string;
  owner_name_ko: string;
  owner_name_en: string;
  registration_num: string;
  mail_num: string;
  address_ko: string;
  address_en: string;
  phone: string;
  fax: string;
  privacy_officer: string;
  admin_email: string;
};